Modification des services mock pour utiliser Long au lieu de Integer pour les IDs

+++ src/test/java/org/romain/demo2/mock/MockProductService.java
@@ -10,7 +10,7 @@ class MockProductService extends ProductService {
     }
 
     @Override
-    public Product toggleAvailability(Integer id) {
+    public Product toggleAvailability(Long id) {
         Product p = new Product();
         p.setId(id);
         p.setAvailable(true); // valeur fictive pour ton test
